requirements.txt
tensorflow
scikit-learn
numpy
matplotlib
seaborn
pandas
streamlit
librosa
pillow
ultralytics